+++ packages/stacked/example/lib/ui/stream_view/stream_counter_viewmodel.dart
@@ -1,4 +1,4 @@
-import 'package:new_architecture/app/locator.dart';
+import 'package:new_architecture/app/app.locator.dart';
 import 'package:new_architecture/services/epoch_service.dart';
 import 'package:stacked/stacked.dart';