+++ packages/stacked_generator/lib/src/generators/router/custom_transition_builder.dart
@@ -1,0 +1,8 @@
+// holds the name and path of the passed
+// transition builder function
+class CustomTransitionBuilder {
+  String name;
+  String import;
+
+  CustomTransitionBuilder(this.name, this.import);
+}