+++ packages/stacked_services/lib/src/dialog/platform_dialog.dart
@@ -10,6 +10,8 @@ class PlatformButton extends StatelessWidget {
   final String text;
   final Function onPressed;
   final bool isCancelButton;
+  final Color confirmationBtnColor;
+  final Color cancelBtnColor;
 
   const PlatformButton({
     Key key,
@@ -17,25 +19,38 @@ class PlatformButton extends StatelessWidget {
     this.isCancelButton = false,
     @required this.text,
     @required this.onPressed,
+    this.confirmationBtnColor,
+    this.cancelBtnColor,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+    // final isDarkMode = Theme.of(context).brightness == Brightness.dark;
     switch (dialogPlatform) {
       case DialogPlatform.Cupertino:
         return CupertinoDialogAction(
-          child: Text(
-            text,
-            style: isCancelButton ? _cancelTextStyle : null,
-          ),
+          child: Text(text,
+              style: isCancelButton
+                  ? cancelBtnColor != null
+                      ? TextStyle(color: cancelBtnColor)
+                      : _cancelTextStyle
+                  : confirmationBtnColor != null
+                      ? TextStyle(color: confirmationBtnColor)
+                      : null),
           onPressed: onPressed,
         );
 
       case DialogPlatform.Material:
       default:
         return FlatButton(
-          child: Text(text, style: isCancelButton ? _cancelTextStyle : null),
+          child: Text(text,
+              style: isCancelButton
+                  ? cancelBtnColor != null
+                      ? TextStyle(color: cancelBtnColor)
+                      : _cancelTextStyle
+                  : confirmationBtnColor != null
+                      ? TextStyle(color: confirmationBtnColor)
+                      : null),
           onPressed: onPressed,
         );
     }
@@ -94,32 +109,37 @@ class PlatformDialog extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+    // final isDarkMode = Theme.of(context).brightness == Brightness.dark;
     switch (dialogPlatform) {
       case DialogPlatform.Cupertino:
         return CupertinoAlertDialog(
-          title: Text(
-            title,
-          ),
-          content: Text(
-            content,
-          ),
+          title: title != null
+              ? Text(
+                  title,
+                )
+              : null,
+          content: content != null
+              ? Text(
+                  content,
+                )
+              : null,
           actions: actions,
         );
       case DialogPlatform.Material:
       default: // TODO: When custom dialog registrations are implemented it'll be shown here
         return AlertDialog(
-          backgroundColor: isDarkMode ? Colors.grey[850] : Colors.white,
           titleTextStyle: Theme.of(context).dialogTheme.titleTextStyle,
           contentTextStyle: Theme.of(context).dialogTheme.contentTextStyle,
-          title: Text(
-            title,
-            style: TextStyle(color: isDarkMode ? Colors.white : Colors.black),
-          ),
-          content: Text(
-            content,
-            style: TextStyle(color: isDarkMode ? Colors.white : Colors.black),
-          ),
+          title: title != null
+              ? Text(
+                  title,
+                )
+              : null,
+          content: content != null
+              ? Text(
+                  content,
+                )
+              : null,
           actions: actions,
         );
     }