+++ packages/stacked_services/lib/stacked_services.dart
@@ -7,3 +7,4 @@ export 'src/snackbar/snackbar_service.dart';
 export 'src/snackbar/snackbar_config.dart';
 export 'src/models/overlay_request.dart';
 export 'src/models/overlay_response.dart';
+export 'src/stacked_service.dart';