+++ packages/stacked/CHANGELOG.md
@@ -1,3 +1,29 @@
+## 1.9.2
+
+### MAJOR UPDATE: FORMS!!!
+
+- Adds brand new form first party support and form generation functionality
+
+## 1.9.1+2
+
+- Fix null error from locator
+
+## 1.9.1+1
+
+- Adds export for `StackedLocator`
+
+## 1.9.1
+
+- Adds `StackedLocator` to remove the apps dependency from GetIt. It's a thin wrapper around the `GetIt` object implementation
+
+## 1.9.0
+
+Big Update! This version of stacked aims to reduce the boiler plate required to setup a stacked application. We introduce the annotation that are used by the stacked_generator to generate the router and locator setup functionality.
+
+## 1.7.7
+
+- Fixes [Multiple Streams ViewModel causing memory overflow](https://github.com/FilledStacks/stacked/issues/190) bug
+
 ## 1.7.6
 
 - Adds `fireOnce` properties to the `ViewModelBuilderWidget`