+++ packages/stacked_themes/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.3
+
+- Use get_it newInstance to avoid clashing with the app
+
+## 0.2.2+2
+
+- Updates packages to the latest
+
 ## 0.2.2+1
 
 - Updates get it to latest